add remove datasets component, add required state, adjust submit
