+++ src/pages/dar_application/ProgressReportApplication.tsx
@@ -1,12 +1,22 @@
-import React from 'react';
-import {DataAccessRequest} from 'src/types/model';
+import React, {useState} from 'react';
+import {DataAccessRequest, Dataset} from 'src/types/model';
 import SubmitProgressReport from '../progress_reports/SubmitProgressReport';
+import SelectableDatasets from '../../pages/dar_application/SelectableDatasets';
 
 type ProgressReportApplicationProps = {
-    dar?: DataAccessRequest, // Dar will be empty if this is a new application
+    dar?: DataAccessRequest, // Dar will be empty if this is an application
+    parentDar?: DataAccessRequest, // Dar will be empty if this is view only
+    datasets: Dataset[],
     readOnlyMode?: boolean
 };
-export const ProgressReportApplication = ({dar, readOnlyMode=true}: ProgressReportApplicationProps) => {
+
+interface FormStateInterface {
+    datasetIds: number[];
+}
+
+export const ProgressReportApplication = ({dar, parentDar, datasets, readOnlyMode=true}: ProgressReportApplicationProps) => {
+    const [formState, setFormState] = useState<FormStateInterface>({datasetIds: []});
+
     return (
         <div className={readOnlyMode ? 'accordion-step-container' : 'step-container'}>
             {!readOnlyMode && <h3>Submit a progress report</h3>}
@@ -19,10 +29,23 @@ export const ProgressReportApplication = ({dar, readOnlyMode=true}: ProgressRepo
                     <h4>Intellectual Property Summary</h4>
                     {dar?.intellectualPropertySummary ?? "PLACEHOLDER Intellectual Property Summary"}
                 </div>
-                {dar?.parentId && <div>
+                <div data-cy='remove-datasets'>
+                    <div className='progress-report-step-card'>
+                        <h2>Step 3: Dataset(s) in this DAR</h2>
+                        <p style={{ marginBottom: '1rem' }}>Currently selected datasets:</p>
+                        <SelectableDatasets
+                            disabled={readOnlyMode}
+                            datasets={datasets}
+                            setSelectedDatasets={(selectedDatasets) => {
+                                setFormState({...formState, datasetIds: selectedDatasets.map((ds) => ds.datasetId)})}
+                            }
+                        />
+                    </div>
+                </div>
+                {!readOnlyMode && parentDar && <div>
                   <SubmitProgressReport
                       progressReport={dar}
-                      parentId={dar.parentId}
+                      parentId={parentDar.referenceId}
                       onSuccess={() => {
                       }}
                       onCancel={() => {
